requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
asyncio
slowapi # for rate limiting
#gunicorn # for multiple workers
supervisor
tenacity==8.2.3
